Added project child sets, refactored, fixed beans initialization

+++ DreamTeamProj/src/main/java/ua/nure/cache/controller/ProjectsController.java
@@ -35,7 +35,7 @@ public class ProjectsController {
 	}
 
 	@RequestMapping(value = "/{id}", method = RequestMethod.DELETE)
-	public void deleteProject(final @PathVariable(value = "id") Long id) {
+	public void deleteProject(final @PathVariable(value = "id") int id) {
 	    projectRepository.delete(id);
 	}
