+++ DreamTeamProj/src/main/java/ua/nure/cache/repository/SourceFieldRepository.java
@@ -3,5 +3,5 @@ package ua.nure.cache.repository;
 import org.springframework.data.repository.CrudRepository;
 import ua.nure.cache.entity.SourceField;
 
-public interface SourceFieldRepository extends CrudRepository<SourceField, Long> {
+public interface SourceFieldRepository extends CrudRepository<SourceField, Integer> {
 }
