+++ DreamTeamProj/src/main/java/ua/nure/cache/repository/UserRoleRepository.java
@@ -3,5 +3,5 @@ package ua.nure.cache.repository;
 import org.springframework.data.repository.CrudRepository;
 import ua.nure.cache.entity.UserRole;
 
-public interface UserRoleRepository extends CrudRepository<UserRole, Long> {
+public interface UserRoleRepository extends CrudRepository<UserRole, Integer> {
 }
